Added Tank class skeleton.
Wow this is going to be a long journey. I never understood why people say "It's about the journey not the destination." Why can't it just be about both? Why can't my work be both about finishing this thing and learning from it? Maybe this is a bit pedantic for a first commit.